requirements.txt
feedparser>=6.0.0
requests>=2.0.0
PyYAML>=6.0.0
